Refactor Directories
+ Renamed ``src/networking`` to ``src/wrapper`` since it makes more sense
+ Moved ``api/base_nodes.py`` to ``api/instance/`
+ Moved ``api/instance_states.py`` to ``api/instance/`
+ Moved ``api/node_scene.py`` to ``api/instance/``

+ Moved ``api/internal/websocket_handler.py`` to ``api/web/manager/``

+ Renamed CommandRouter to BaseMessageRouter
+ Renamed ``api/web/command_router.py`` to ``api/web/message_router.py``

+ Moved ``api/actions/`` to ``api/instance/actions/``

+ Moved ``api/internal/websocket_manager.py`` to ``api/`
+ Moved ``api/server_instance.py`` to ``api/instance/server_instance.py``
