+++ src/nodeserver/api/instance/base_nodes.py
@@ -1,8 +1,8 @@
 
 from typing import Any
 
-from nodeserver.networking.nodes.data.node_data_types import SuperSlotTypes
-from nodeserver.networking.nodes.node.base_nodes import NodeMirror, SlotMirror
+from nodeserver.wrapper.nodes.data.node_data_types import SuperSlotTypes
+from nodeserver.wrapper.nodes.node.base_nodes import NodeMirror, SlotMirror
 
 
 # TODO: Change node inputs to a NodeOutput class or NodeInputs class
