+++ src/nodeserver/api/internal/internal_protocols.py
@@ -1,7 +1,7 @@
 from typing import Protocol
 
-from nodeserver.api.node_scene import NodeScene
-from nodeserver.networking.nodes.helpers.scene_manager import MirrorSceneManager
+from nodeserver.api.instance.node_scene import NodeScene
+from nodeserver.wrapper.nodes.helpers.scene_manager import MirrorSceneManager
 
 class InstanceProtocol(Protocol):
     mirror_manager: MirrorSceneManager
