+++ src/nodeserver/api/web/message_router.py
@@ -2,15 +2,15 @@ from enum import Enum
 import json
 import logging
 
-from nodeserver.api.actions.action_controller import Action
-from nodeserver.api.internal.websocket_messages import ClientMessageWrapper
+from nodeserver.api.instance.actions.action_controller import Action
+from nodeserver.api.web.websocket_messages import ClientMessageWrapper
 from nodeserver.api.web.requests.client_requests import ClientCommand, MsgConnectionAction, MsgInstanceCommand, MsgInstanceState, MsgLoadScene, MsgLoopState, MsgNodeAction, MsgSimple
 from nodeserver.api.web.websocket_protocol import ClientMessages, ServerMessages
-from nodeserver.api.server_instance import ServerInstance
-from nodeserver.networking.nodes.helpers.file.node_scene_dataclasses import ConnectionSceneData, NodeSceneData
+from nodeserver.api.instance.server_instance import ServerInstance
+from nodeserver.wrapper.nodes.helpers.file.node_scene_dataclasses import ConnectionSceneData, NodeSceneData
 
 COMMAND_LOGGER = logging.getLogger("nds.commands")
-class CommandRouter:
+class BaseMessagerouter:
     def route_message(self, message: ClientMessageWrapper, instance: ServerInstance) -> dict | None:
         COMMAND_LOGGER.info(f"Routing command: {message.msg.type}")
