+++ src/nodeserver/wrapper/nodes/helpers/file/typing_file_reader.py
@@ -3,14 +3,14 @@ from dataclasses import dataclass
 import json
 from typing import Callable
 
-from nodeserver.api.base_nodes import BaseNode
-from nodeserver.networking.nodes.data.custom_data_types import CustomSlotType
-from nodeserver.networking.nodes.data.node_data import NodeData
-from nodeserver.networking.nodes.data.node_data_types import BaseSlotType
-from nodeserver.networking.nodes.helpers.file.node_scene_dataclasses import SceneData
-from nodeserver.networking.nodes.helpers.file.type_dataclasses import NodeTypeData, SlotData, SlotTypeData, TypeFile
-from nodeserver.networking.nodes.helpers.node_constructor import BaseMirrorConstructor, CustomMirrorConstructor
-from nodeserver.networking.nodes.node.base_nodes import NodeMirror
+from nodeserver.api.instance.base_nodes import BaseNode
+from nodeserver.wrapper.nodes.data.custom_data_types import CustomSlotType
+from nodeserver.wrapper.nodes.data.node_data import NodeData
+from nodeserver.wrapper.nodes.data.node_data_types import BaseSlotType
+from nodeserver.wrapper.nodes.helpers.file.node_scene_dataclasses import SceneData
+from nodeserver.wrapper.nodes.helpers.file.type_dataclasses import NodeTypeData, SlotData, SlotTypeData, TypeFile
+from nodeserver.wrapper.nodes.helpers.node_constructor import BaseMirrorConstructor, CustomMirrorConstructor
+from nodeserver.wrapper.nodes.node.base_nodes import NodeMirror
 
 class TypeFileReader:
     _node_types_version: int = -1
