+++ src/nodeserver/wrapper/nodes/node/node_utils.py
@@ -1,8 +1,8 @@
 from collections import deque
 
-from nodeserver.api.base_nodes import BaseNode
-from nodeserver.networking.nodes.data.node_data_types import SuperSlotTypes
-from nodeserver.networking.nodes.node.base_nodes import NodeMirror
+from nodeserver.api.instance.base_nodes import BaseNode
+from nodeserver.wrapper.nodes.data.node_data_types import SuperSlotTypes
+from nodeserver.wrapper.nodes.node.base_nodes import NodeMirror
 
 class NodeUtils:
     @staticmethod
