+++ tests/scene_file_test.py
@@ -1,5 +1,5 @@
-from nodeserver.networking.nodes.helpers.file.node_scene_reader import SceneFileReader
-from nodeserver.networking.nodes.helpers.file.typing_file_reader import TypeFileReader
+from nodeserver.wrapper.nodes.helpers.file.node_scene_reader import SceneFileReader
+from nodeserver.wrapper.nodes.helpers.file.typing_file_reader import TypeFileReader
 from test_data import SCENE_DATA_JSON, TYPE_FILE_JSON
 
 types_file = TypeFileReader()
