+++ tests/typing_file_test.py
@@ -1,10 +1,10 @@
 import logging
 
-from nodeserver.api.base_nodes import BaseNode
-from nodeserver.networking.nodes.data.node_data import NodeData
-from nodeserver.networking.nodes.data.node_data_types import BaseNodeType
-from nodeserver.networking.nodes.helpers.file.typing_file_reader import TypeFileReader
-from nodeserver.networking.utils.uuid_utils import IDGenerator
+from nodeserver.api.instance.base_nodes import BaseNode
+from nodeserver.wrapper.nodes.data.node_data import NodeData
+from nodeserver.wrapper.nodes.data.node_data_types import BaseNodeType
+from nodeserver.wrapper.nodes.helpers.file.typing_file_reader import TypeFileReader
+from nodeserver.wrapper.utils.uuid_utils import IDGenerator
 from test_data import TYPE_FILE_JSON
 
 LOGGER = logging.Logger("Logger")
